stranger_demo_bootstrap/requirements-demo.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
streamlit==1.44.1
pydantic==2.11.3
pyyaml==6.0.2
numpy==1.26.4
pandas==2.2.3
scipy==1.15.2
opencv-python==4.11.0.86
Pillow==11.1.0
tqdm==4.67.1
sqlalchemy==2.0.40
alembic==1.15.2
python-multipart==0.0.20
onnxruntime==1.21.0
ultralytics==8.3.99
lapx==0.5.11
supervision==0.25.1
filterpy==1.4.5
networkx==3.4.2
httpx==0.28.1
orjson==3.10.16
pytest==8.3.5
